+++ facebook_auth_desktop/example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import facebook_auth_desktop
+import flutter_secure_storage_macos
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FacebookAuthDesktopPlugin.register(with: registry.registrar(forPlugin: "FacebookAuthDesktopPlugin"))
+  FlutterSecureStorageMacosPlugin.register(with: registry.registrar(forPlugin: "FlutterSecureStorageMacosPlugin"))
+}