+++ facebook_auth_desktop/lib/src/custom_http_client.dart
@@ -1,0 +1,10 @@
+// coverage:ignore-file
+
+import 'package:http/http.dart' as http;
+import 'package:http/http.dart';
+
+class CustomHttpClient {
+  Future<Response> get(Uri url) async {
+    return http.get(url);
+  }
+}