+++ facebook_auth_web/lib/src/facebook_auth_plugin.dart
@@ -135,7 +135,7 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   ///
   /// calls the FB.init interop
   @override
-  Future<void> webInitialize({
+  Future<void> webAndDesktopInitialize({
     required String appId,
     required bool cookie,
     required bool xfbml,
@@ -214,11 +214,16 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
               }
             }
             c.complete(
-              FacebookPermissions(granted: granted, declined: declined),
+              FacebookPermissions(
+                granted: granted,
+                declined: declined,
+              ),
             );
           } on PlatformException catch (e) {
             print(
-              StackTrace.fromString(e.message ?? 'unknown error'),
+              StackTrace.fromString(
+                e.message ?? 'unknown error',
+              ),
             );
             c.complete(null);
           }
@@ -254,13 +259,17 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
       final String? status = response['status'];
 
       if (status == null) {
-        return LoginResult(status: LoginStatus.failed);
+        return LoginResult(
+          status: LoginStatus.failed,
+        );
       }
       if (status == 'connected') {
         final Map<String, dynamic> authResponse = response['authResponse'];
 
         final expires = DateTime.now().add(
-          Duration(seconds: authResponse['expiresIn']),
+          Duration(
+            seconds: authResponse['expiresIn'],
+          ),
         );
 
         // create a Login Result with an accessToken
@@ -286,9 +295,15 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
       } else if (status == 'unknown') {
         return LoginResult(status: LoginStatus.cancelled);
       }
-      return LoginResult(status: LoginStatus.failed, message: 'unknown error');
+      return LoginResult(
+        status: LoginStatus.failed,
+        message: 'unknown error',
+      );
     } on PlatformException catch (e) {
-      return LoginResult(status: LoginStatus.failed, message: e.message);
+      return LoginResult(
+        status: LoginStatus.failed,
+        message: e.message,
+      );
     }
   }
 
@@ -296,7 +311,9 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   void _checkResponseError(Map<String, dynamic> response) {
     if (response['error'] != null) {
       throw PlatformException(
-          code: "REQUEST_ERROR", message: response['error']['message']);
+        code: "REQUEST_ERROR",
+        message: response['error']['message'],
+      );
     }
   }
 
@@ -304,7 +321,9 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   bool get isWebSdkInitialized => _initialized;
 
   @override
-  Future<void> autoLogAppEventsEnabled(bool enabled) async {}
+  Future<void> autoLogAppEventsEnabled(bool enabled) {
+    throw UnimplementedError();
+  }
 
   @override
   Future<bool> get isAutoLogAppEventsEnabled => Future.value(false);